requirements.txt
# Telegram bot framework
python-telegram-bot==20.7

# HTTP requests (for API login)
requests==2.31.0